car-recsys-system/etl/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pandas==2.1.4
numpy==1.26.3

# Search & Vector
elasticsearch==8.11.1
qdrant-client==1.7.0

# Machine Learning
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
scikit-learn==1.4.0

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
